#91 design: dashboard records 페이지 재분리 작업 및 dashboard UI 작업 수정예정
