+++ app/dashboard/page.tsx
@@ -1,281 +1,112 @@
-"use client";
-
-import { useState } from "react";
-import { useEffect } from "react";
-import axios from "axios";
-import styles from "./records.module.scss";
 import Image from "next/image";
-import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "@/store/store";
-import bal from "@/public/blueArrow_L.png";
-import bar from "@/public/blueArrow_R.png";
-import gal from "@/public/grayArrow_L.png";
-import gar from "@/public/grayArrow_R.png";
-import Lottie from "react-lottie-player";
-import TrophyAnim from "@/public/animation_trophy.json";
-import FailAnim from "@/public/animation_fail.json";
-import CheckAnim from "@/public/animation_greencheck2.json";
-import Doughnut from "./doughnut";
-import { averageSave } from "@/slices/chartInfo";
-import { onDashboard } from "@/util/onDashboard";
+import React from "react";
+import styles from "./dashboard.module.scss";
+import ads_banner1 from "../../public/ads_banner1.png";
+import newsIcon from "../../public/Icon_newspaper.png";
+import newsImg from "../../public/news_img.png";
+import student1 from "../../public/student1.png";
 
-function Trophy() {
+const Dashboard = () => {
   return (
-    <Lottie
-      loop={false}
-      animationData={TrophyAnim}
-      play
-      style={{ margin: "0 auto", width: "70%", height: "auto" }}
-    />
-  );
-}
-
-function Failed() {
-  return (
-    <Lottie
-      loop={false}
-      animationData={FailAnim}
-      play
-      style={{ margin: "0 auto", width: "70%", height: "auto" }}
-    />
-  );
-}
-
-function Check() {
-  return (
-    <Lottie
-      loop={false}
-      animationData={CheckAnim}
-      play
-      style={{ margin: "0 auto", width: "70%", height: "auto" }}
-    />
-  );
-}
-export interface userRecord {
-  title: string;
-  challengeTime: number;
-  date: string;
-  challengeStatus: string;
-  description: string;
-  difficulty: number;
-}
-
-// 1. 먼저 userChallenges 에서 challengeTime 을 가져와 challengeTimeAvg(평균을)를 낸다
-// state 를 하나 새로 파서 map 함수로 index를 순회할 때마다 이 State에 챌린지 타임 퍼센테이지 값을 계속 더해줌
-// 그렇게 다 더해진 값을 userChallenges 의 길이로 나누면 평균이 나옴 -> 소숫점 반올림
-// 그러면 이 값을 가지고 다음 과정으로 진행
-// 2. 이 평균 값을 도넛차트 컴포넌트에 props 로 전달함?
-// 3. 그러며는 이 전달한 값으로 도넛의 data 를 바꿔주기만 하면 댐
-
-export default function Records() {
-  interface chartState {
-    timeAvrg: Number;
-    diffAvrg: Number;
-    succeedCount: Number;
-  }
-  const initialChartState = {
-    timeAvrg: 0,
-    diffAvrg: 0,
-    succeedCount: 0,
-  };
-  let [userChallenges, setUserChallenges] = useState<userRecord[]>([]);
-  let [curIndex, setCurIndex] = useState(0);
-  let [chartProps, setChartProps] = useState(initialChartState);
-  const userInfo = useSelector((state: RootState) => state.user);
-  const chartInfo = useSelector((state: RootState) => state.chart);
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    onDashboard()
-      .then((data) => {
-        setUserChallenges([...data]);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, []);
-
-  useEffect(() => {
-    let copyTime: any = [];
-    userChallenges.map((a, i) => {
-      copyTime.push(a.challengeTime);
-    });
-    const timeAvrg =
-      copyTime.reduce((p: any, c: any) => p + c, 0) / copyTime.length;
-    let copyDiff: any = [];
-    userChallenges.map((a, i) => {
-      copyDiff.push(a.difficulty);
-    });
-    const diffAvrg =
-      copyDiff.reduce((p: any, c: any) => p + c, 0) / copyDiff.length;
-    let copySucceed: any = [];
-    userChallenges.map((a, i) => {
-      if (a.challengeStatus === "succeed") {
-        copySucceed.push(a.challengeStatus);
-      }
-    });
-    const succeedCount = (copySucceed.length / userChallenges.length) * 100;
-    let newChartProps = {
-      ...chartProps,
-      timeAvrg: timeAvrg,
-      diffAvrg: diffAvrg,
-      succeedCount: succeedCount,
-    };
-    setChartProps(newChartProps);
-  }, [userChallenges]);
-
-  useEffect(() => {
-    dispatch(averageSave(chartProps));
-  }, [chartProps, dispatch]);
-
-  // 시간 퍼센테이지 평균값 계산 및 Dispatch
-  // const avgTimeFunc = () => {
-  //   let copy: any = [];
-  //   userChallenges.map((a, i) => {
-  //     copy.push(a.challengeTime);
-  //   });
-  //   const average = copy.reduce((p: any, c: any) => p + c, 0) / copy.length;
-  //   console.log(average);
-  //   dispatch(timeSave(average));
-  // };
-
-  // // 난이도 평균값 계산 및 가공 후 Dispatch
-  // const diffValueFunc = () => {
-  //   let copy: any = [];
-  //   userChallenges.map((a, i) => {
-  //     copy.push(a.difficulty);
-  //   });
-  //   const diffAvg = copy.reduce((p: any, c: any) => p + c, 0) / copy.length;
-  //   dispatch(diffSave(diffAvg));
-  // };
-
-  // 성공횟수 카운트 및 퍼센테이지 가공 후 Dispatch
-
-  const cardImg = (challengeItem: any) => {
-    // cardImg의 파라미터로 challenge를 넘겨받으면 굳이 전체 게시글 중 인덱스로 탐색할 필요 없이
-    // challengeItem 안의 challengeStatus를 참조해 cardImg를 return해줄 수 있음!
-    if (challengeItem.challengeStatus === "succeed") {
-      return <Trophy />;
-    } else if (challengeItem.challengeStatus === "failed") {
-      return <Failed />;
-    } else {
-      return <Check />;
-    }
-  };
-
-  const diffReturn = () => {
-    if (chartProps.diffAvrg >= 0 && chartProps.diffAvrg < 25) {
-      return <h3>유망주</h3>;
-    } else if (chartProps.diffAvrg > 25 && chartProps.diffAvrg < 50) {
-      return <h3>에이스</h3>;
-    } else if (chartProps.diffAvrg > 50 && chartProps.diffAvrg < 75) {
-      return <h3>챌린저</h3>;
-    } else if (chartProps.diffAvrg > 75 && chartProps.diffAvrg <= 100) {
-      return <h3>챔피언</h3>;
-    }
-  };
-
-  return (
-    <div className={styles.ground}>
-      <div className={styles.titleSec}>
-        <h3>원아워 레코즈</h3>
-        <p>{userInfo.name}님의 최근 원아워 레코즈를 정리해드릴게요</p>
+    <>
+      <div className={styles.adsSection}>
+        <Image
+          className={styles.ads_banner1}
+          src={ads_banner1}
+          width={1620}
+          alt="ads1-nike"
+        />
+        <button>자세히</button>
       </div>
-      <div className={styles.listUpper}>
-        {userChallenges ? (
-          <>
-            <Image
-              src={curIndex <= 0 ? gal : bal}
-              alt="arrow"
-              height={50}
-              onClick={() => {
-                setCurIndex((curIndex) => curIndex - 1); // 왼쪽 버튼 클릭 시 지금보다 인덱스가 줄어들어야 함!
-                // 인덱스가 0보다 작거나 같아지면 인덱스를 0으로 고정
-                if (curIndex <= 0) setCurIndex(0);
-              }}
-              style={{ cursor: "pointer" }}
-            />
-            {userChallenges
-              .slice(curIndex, curIndex + 3)
-              .map((challengeItem, i) => {
-                // a -> challengeItem으로 변경해 알아보기 쉽게 했어요
-                return (
-                  <div key={i}>
-                    <div className={styles.recordsBox} key={`records${i}`}>
-                      {/* 챌린지 아이템을 cardImg 파라미터로 넘겨주게 함!! 위에 cardImg 함수 수정했어요  */}
-                      <h4 className={styles.cardTitle}>
-                        {challengeItem.title}
-                      </h4>
-                      <p className={styles.cardDate}>{challengeItem.date}</p>
-                      <div className={styles.cardImg}>
-                        {cardImg(challengeItem)}
-                      </div>
-                      <p className={styles.cardDesc}>
-                        &quot; {challengeItem.description} &quot;
-                      </p>
-                      <div className={styles.progressWrap}>
-                        <div className={styles.progressTitle}>
-                          <span>투자한 시간</span>
-                          <span>( 60분 기준 )</span>
-                        </div>
-                        <div className={styles.progressBar}>
-                          <div
-                            className={styles.progressState}
-                            style={{ width: `${challengeItem.challengeTime}%` }}
-                          />
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                );
-              })}
-            <Image
-              src={curIndex >= userChallenges.length - 3 ? gar : bar}
-              alt="arrow"
-              height={50}
-              onClick={() => {
-                setCurIndex((curIndex) => curIndex + 1); // 오른쪽 버튼 클릭 시 지금보다 인덱스가 커져야 함
-                if (curIndex >= userChallenges.length - 3) {
-                  // 인덱스가 전체 챌린지 개수-3보다 같거나 커지면 인덱스를 전체 챌린지 개수-3으로 고정
-                  setCurIndex(userChallenges.length - 3);
-                }
-              }}
-              style={{ cursor: "pointer" }}
-            />
-          </>
-        ) : (
-          ""
-        )}
-      </div>
-      <div className={styles.titleSec2}>
-        <h3>#불타는 열정</h3>
-        <p>홍당무님의 기록을 기반으로 목표 실천율을 분석해봤어요</p>
-      </div>
-      <div className={styles.chartUpper}>
-        <div className={styles.donutSec}>
-          <h4>시간 계획</h4>
-          <h2 className={styles.timeAvrg}>
-            {Math.round(chartProps.timeAvrg)}%
-          </h2>
-          <Doughnut value={chartInfo.challengeTimeAvg} id={"timeAvrgChart"} />
-          <p>전체 집계 대비 수행시간 정보</p>
-        </div>
-        <div className={styles.donutSec}>
-          <h4>도전 난이도</h4>
-          {diffReturn()}
-          <Doughnut value={chartInfo.difficultAvg} id={"diffAvrgChart"} />
-          <p>평균 도전 난이도 정보</p>
-        </div>
-        <div className={styles.donutSec}>
-          <h4>목표 달성도</h4>
-          <h2 className={styles.succeedCnt}>
-            {Math.round(chartProps.succeedCount)}%
-          </h2>
-          <Doughnut value={chartInfo.succeedCount} id={"succeedRatioChart"} />
-          <p>전체 집계 대비 90% 이상 달성 건</p>
+      <div className={styles.people_news}>
+        <h2>원아워 피플뉴스</h2>
+        <Image className={styles.newsIcon} src={newsIcon} alt="news-icon" />
+        <div className={styles.newsBox}>
+          <div className={styles.newsImage}>
+            <Image className={styles.news_pic} src={newsImg} alt="news-image" />
+          </div>
+          <div className={styles.newsText}>
+            <h3>
+              한국의 금융을 새로 쓰는 활동가, 김태훈 (주)뱅크샐러드 대표의
+              시작점
+            </h3>
+            <p>
+              처음 회사를 설립할 때에 제가 바라본 목표는 “내 자산관리를
+              어떻게하면 잘 할수 있을까?” 였어요
+              <br />
+              금융에 대해서 고민하던 저는, 사업을 시작하기 전에 필요한
+              제반사항이 바로 제가 직접 경험하는 것 이었죠
+              <br />
+              그래서 하루에 무조건 3번 자산 기록하기부터 시작됐어요 제 실천이.
+              그게 지금의 뱅크샐러드 가족들을
+              <br />
+              만들어 준 것 같아요. 여러분도 자신만의 목표를 열심히 실천하셔서
+              성장하셨으면 좋겠습니다!
+            </p>
+            <button>본문으로</button>
+          </div>
         </div>
       </div>
-      <div className={styles.footer}></div>
-    </div>
+      <div className={styles.weekly_challenge}>
+        <h2>WEEKLY CHALLENGE</h2>
+        <h4>
+          원아워는 수험생에게 아주 유용한 서비스인 것 같아요, 수험생의 원아워
+          사용법 꿀팁 들어갑니다 ! 🍯
+        </h4>
+        <p className={styles.subText}>
+          딱딱한 타이머 공부법 대신 과목별로 나누어 챌린지를 수행해보세요.
+          트로피도 얻고, 성취감도 더욱 느낄 수 있는 것 같습니다!
+        </p>
+        <p>
+          원아워를 시작한 계기도 ‘꾸준함’ 을 기르기 위해서였습니다. 카드를
+          하나하나씩 채울때마다 오늘 해야할 공부량이
+          <br />
+          가시적으로 들어오기도 하고 한시간이라는 적절한 분량이 주는 집중도는 제
+          기대보다 훨씬 높았어요.
+          <br />
+          물론 난이도를 선정할 때 주관적인 부분이 개입될 수 밖에 없기 때문에
+          원아워 레코드 지표가
+          <br />
+          알맞은 결과인가? 라는 생각도 들었었지만, 저는 제가 어려워하던 과목을
+          공부할 때 난이도를
+          <br />
+          높은 쪽으로 설정해두었고, 챌린지 제목도 단순히 수학, 과학과 같이
+          포괄적인 과목명이 아닌
+          <br />
+          “화학2 - 엔트로피의 성질과 열역학” 처럼 세부적인 이론을 쪼개가며
+          기록했어요.
+          <br />
+          이렇게 하니 제가 두려워하는 과목과 이론이 정리가 되었고 그 부분을
+          중심으로
+          <br />
+          취약점을 보완하기 더욱 쉬워졌던 것 같습니다. 저처럼 세부 과목을 쪼개는
+          것도
+          <br />
+          방법이 될 수 있고, 한 시간 목표를 연속으로 수행할수도 있으니 이를 잘
+          이용하여
+          <br />
+          본인만의 공부전략을 세워보시기를 ‘적극’ 추천드립니다. 체계적인 습관을
+          들이는
+          <br />
+          가장 쉬운 방법이 꾸준함이라고 생각합니다. 원아워의 기능은 정말
+          단순하면서도
+          <br />
+          꾸준함을 기르는 데에 최적화된 서비스라고 생각해요. 얼마 남지않은
+          수학능력시험
+          <br />
+          모두 잘 준비하셔서 좋은결과 꼭! 이뤄내시길 바랄게요. 화이팅! 대한민국
+          수험생!
+        </p>
+        <Image className={styles.student1} src={student1} alt="student1" />
+        <p className={styles.student_info}>
+          숭실고등학교 3학년
+          <br />
+          성의진 학생
+        </p>
+        <div className={styles.boxMargin}></div>
+      </div>
+    </>
   );
-}
+};
+
+export default Dashboard;
